Rename QueryPageServices => Services
Ticket: #260
