+++ src/app/Main/Body/QueryPage/Services/PathServices/generatePathString.service.ts
