+++ src/app/Main/Body/QueryPage/Services/PathServices/getFilterFromPath.service.ts
