+++ src/app/Main/Body/QueryPage/Services/TableServices/calculateRelativeTableData.service.ts
