+++ src/app/Main/Body/QueryPage/Services/TableServices/generateStatisticTableDataAbs.service.ts
