+++ src/app/Main/Body/QueryPage/Services/adaptCountedIsolatesGroups.service.ts
