+++ src/app/Main/Body/QueryPage/Services/adaptFilterFromAPI.service.ts
