+++ src/app/Main/Body/QueryPage/Services/checkIfFilterIsSet.service.ts
